Add dry run functionality [skip travis]

+++ update.sh
@@ -1,16 +1,31 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
-supported_tag="$1"
-echo "supported_tag: $supported_tag"
+dry_run=false
 
-if [[ -d "$supported_tag" ]]; then
-  echo "Already supported tag: $supported_tag"
+while getopts ":t:d" opt; do
+  case $opt in
+  t)
+    new_supported_tag="$OPTARG"
+    ;;
+  d)
+    dry_run=true
+    dry_run_dir=dry-run
+    ;;
+  \?)
+    echo "Invalid option: -$OPTARG" >&2
+    ;;
+  esac
+done
+echo "new_supported_tag:$new_supported_tag, dry_run:$dry_run"
+
+if [[ -d "$new_supported_tag" && "$dry_run" == false ]]; then
+  echo "Already supported tag: $new_supported_tag"
   exit 1
 fi
 
-version="${supported_tag%%/*}"
-variant="$(basename "$supported_tag")"
+version="${new_supported_tag%%/*}"
+variant="$(basename "$new_supported_tag")"
 java_variant="${variant%%-*}"
 
 if [[ "$java_variant" != jdk* ]]; then
@@ -20,7 +35,7 @@ fi
 
 os_variant="${variant#$java_variant-}"
 
-if [[ "${os_variant}" == "${java_variant}" ]] ; then
+if [[ "${os_variant}" == "${java_variant}" ]]; then
   os_variant=
 fi
 
@@ -34,21 +49,25 @@ base_image="openjdk:${java_variant:3}-${java_variant:0:3}${os_variant:+-$os_vari
 echo "base_image: $base_image"
 
 full_version=$(
-  curl -fsSL --compressed "https://releases.liferay.com/portal/$version/" \
-    | grep -E '<a href="liferay-ce-portal-tomcat-'"$version"'-[0-9]+.tar.gz.MD5' \
-    | sed -r 's!.*<a href="liferay-ce-portal-tomcat-([^"/]+)/?.tar.gz.MD5".*!\1!'
+  curl -fsSL --compressed "https://releases.liferay.com/portal/$version/" |
+    grep -E '<a href="liferay-ce-portal-tomcat-'"$version"'-[0-9]+.tar.gz.MD5' |
+    sed -r 's!.*<a href="liferay-ce-portal-tomcat-([^"/]+)/?.tar.gz.MD5".*!\1!'
 )
 download_url="https://releases.liferay.com/portal/$version/liferay-ce-portal-tomcat-$full_version.tar.gz"
 echo "download_url: $download_url"
 
 md5=$(
-  curl -fsSL "$download_url.MD5" \
-    | cut -d' ' -f1
+  curl -fsSL "$download_url.MD5" |
+    cut -d' ' -f1
 )
 echo "md5: $md5"
 
 echo "Adding Dockerfile for $version/$variant..."
-mkdir -p "$version/$variant"
+release_dir="$version/$variant"
+if [[ "$dry_run" == true ]]; then
+  release_dir="$dry_run_dir/$release_dir"
+fi
+mkdir -p "$release_dir"
 
 sed \
   -e 's!%%BASE_IMAGE%%!'"$base_image"'!g' \
@@ -56,43 +75,45 @@ sed \
   -e 's!%%LIFERAY_DOWNLOAD_URL%%!'"$download_url"'!g' \
   -e 's!%%LIFERAY_DOWNLOAD_MD5%%!'"$md5"'!g' \
   "Dockerfile${os_variant:+-$os_variant}.template" \
-  > "$version/$variant/Dockerfile"
+  > "$release_dir/Dockerfile"
 
-su_tool='gosu'
-if [[ ${os_variant} == alpine ]] ; then
-  su_tool='su-exec'
+su_tool=gosu
+if [[ ${os_variant} == alpine ]]; then
+  su_tool=su-exec
 fi
 
 sed \
   -e 's!%%SU_TOOL%%!'"$su_tool"'!g' \
   docker-entrypoint.template \
-  > "$version/$variant/docker-entrypoint.sh"
-chmod +x "$version/$variant/docker-entrypoint.sh"
+  > "$release_dir/docker-entrypoint.sh"
+chmod +x "$release_dir/docker-entrypoint.sh"
 
-travis="$(awk '/matrix:/{print;getline;$0="    - VERSION='"$version"' VARIANT='"$variant"'"}1' ./.travis.yml)"
-echo "Modifying .travis.yml with new VERSION/VARIANT[$version/$variant]..."
-echo "$travis" > .travis.yml
+if [[ "$dry_run" == false ]]; then
+  travis="$(awk '/matrix:/{print;getline;$0="    - VERSION='"$version"' VARIANT='"$variant"'"}1' ./.travis.yml)"
+  echo "Modifying .travis.yml with new VERSION/VARIANT[$version/$variant]..."
+  echo "$travis" >.travis.yml
 
-if [[ -f ./supported-tags ]]; then
-  if grep -q "$version" ./supported-tags; then
-    echo "Found in supported-tags: release[$version]"
-    echo "$supported_tag" >> ./supported-tags
+  if [[ -f ./supported-tags ]]; then
+    if grep -q "$version" ./supported-tags; then
+      echo "Found in supported-tags: release[$version]"
+      echo "$new_supported_tag" >>./supported-tags
+    else
+      echo "Not found in supported-tags: release[$version]"
+      echo "$new_supported_tag" >./supported-tags
+
+      for release_dir in $(ls -d [7-9]*/); do
+        if [[ ${new_supported_tag} != "$release_dir"* ]]; then
+          echo "Removing directory: release[$release_dir]..."
+          rm -rf "$release_dir"
+        fi
+      done
+    fi
   else
-    echo "Not found in supported-tags: release[$version]"
-    echo "$supported_tag" > ./supported-tags
-
-    for release_path in $(ls -d [7-9]*/); do
-      if [[ ${supported_tag} != "$release_path"* ]]; then
-        echo "Removing directory: release[$release_path]..."
-        rm -rf "$release_path"
-      fi
-    done
+    echo "Creating supported-tags file..."
+    echo "$new_supported_tag" >./supported-tags
   fi
-else
-  echo "Creating supported-tags file..."
-  echo "$supported_tag" > ./supported-tags
-fi
 
-git add .
-git commit -m "Add supported tag [$supported_tag]"
-git push
+  git add .
+  git commit -m "Add new supported tag [$new_supported_tag]"
+  git push
+fi
